Use a process instead of task for TimerTask
This magically resolved the hang in the Monkey Island II intro.
I still don't know what caused the issue in the Int3 handling there...

Since the timer task is a process now, we are not allowed to just kill it via
DeleteTask(). But then it started failing at exit with a privilege violation.

Turns out, the real fix is taking __interrupt off the TimerTask function.
It caused the function to return with RTE, which is apparently a supervisor opcode.